chore: share curriculum api logic/types for canonical queries
