+++ src/components/QuizContainerNew/QuestionsListNew/QuestionListItemNew/Answers/MCAnswers.tsx
@@ -7,7 +7,7 @@ import Box from "@/components/Box";
 import Flex from "@/components/Flex";
 import Icon from "@/components/Icon";
 import Typography from "@/components/Typography";
-import { MCAnswer } from "@/node-lib/curriculum-api-2023/queries/lessonOverview/lessonOverview.schema";
+import { MCAnswer } from "@/node-lib/curriculum-api-2023/shared.schema";
 
 export const MCAnswers = (props: {
   answers: MCAnswer[];
