+++ src/components/QuizContainerNew/QuestionsListNew/QuestionListItemNew/Answers/MatchAnswers.tsx
@@ -1,12 +1,11 @@
 import { VisuallyHidden } from "react-aria";
 
-import { removeMarkdown } from "../../../quizUtils";
-
+import { removeMarkdown } from "@/components/QuizContainerNew/quizUtils";
 import Box from "@/components/Box";
 import Flex from "@/components/Flex";
 import Icon from "@/components/Icon";
 import Typography from "@/components/Typography";
-import { MatchAnswer } from "@/node-lib/curriculum-api-2023/queries/lessonOverview/lessonOverview.schema";
+import { MatchAnswer } from "@/node-lib/curriculum-api-2023/shared.schema";
 
 export const MatchAnswers = ({
   answers,
