+++ src/components/QuizContainerNew/QuestionsListNew/QuestionListItemNew/Answers/ShortAnswers.tsx
@@ -6,7 +6,7 @@ import Box from "@/components/Box";
 import Flex from "@/components/Flex";
 import Icon from "@/components/Icon";
 import Typography from "@/components/Typography";
-import { ShortAnswer } from "@/node-lib/curriculum-api-2023/queries/lessonOverview/lessonOverview.schema";
+import { ShortAnswer } from "@/node-lib/curriculum-api-2023/shared.schema";
 
 export const ShortAnswers = ({ answers }: { answers: ShortAnswer[] }) => {
   const answerString = answers.reduce((acc, cur) => {
