+++ src/components/QuizContainerNew/QuestionsListNew/QuestionListItemNew/QuizImage.tsx
@@ -1,6 +1,6 @@
 import { FC, useState } from "react";
 
-import { StemImageObject } from "@/node-lib/curriculum-api-2023/queries/lessonOverview/lessonOverview.schema";
+import { StemImageObject } from "@/node-lib/curriculum-api-2023/shared.schema";
 import Flex from "@/components/Flex";
 import OakImage from "@/components/OakImage";
