+++ src/node-lib/curriculum-api-2023/index.ts
@@ -4,7 +4,7 @@ import sdk from "./sdk";
 import lessonOverviewQuery from "./queries/lessonOverview/lessonOverview.query";
 import lessonListingQuery from "./queries/lessonListing/lessonListing.query";
 import subjectListingQuery from "./queries/subjectListing/subjectListing.query";
-import lessonDownloadsQuery from "./queries/downloads/downloads.query";
+import lessonDownloadsQuery from "./queries/lessonDownloads/lessonDownloads.query";
 import programmeListingQuery from "./queries/programmeListing/programmeListing.query";
 import unitListingQuery from "./queries/unitListing/unitListing.query";
 import subjectPhaseOptionsQuery from "./queries/subjectPhaseOptions/subjectPhaseOptions.query";
