+++ src/node-lib/curriculum-api-2023/queries/lessonDownloads/lessonDownloads.query.ts
@@ -2,7 +2,7 @@ import errorReporter from "../../../../common-lib/error-reporter";
 import OakError from "../../../../errors/OakError";
 import { Sdk } from "../../sdk";
 
-import lessonDownloadsSchema from "./downloads.schema";
+import lessonDownloadsSchema from "./lessonDownloads.schema";
 
 const lessonDownloadsQuery =
   (sdk: Sdk) =>
