+++ src/node-lib/curriculum-api-2023/queries/lessonDownloads/lessonDownloads.schema.ts
@@ -1,0 +1,15 @@
+import { z } from "zod";
+
+import {
+  baseLessonDownloadsSchema,
+  lessonPathwaySchema,
+} from "../../shared.schema";
+
+export const lessonDownloadsSchema = z.intersection(
+  baseLessonDownloadsSchema,
+  lessonPathwaySchema,
+);
+
+export type LessonDownloadsPageData = z.infer<typeof lessonDownloadsSchema>;
+
+export default lessonDownloadsSchema;
