+++ src/node-lib/curriculum-api-2023/queries/lessonDownloadsCanonical/lessonDownloadsCanonical.schema.ts
@@ -1,33 +1,12 @@
 import { z } from "zod";
 
-import { pathwaySchema } from "../lessonOverviewCanonical/lessonOverviewCanonical.schema";
+import {
+  baseLessonDownloadsSchema,
+  lessonPathwaySchema,
+} from "../../shared.schema";
 
-export const lessonDownloadsCanonicalSchema = z.object({
-  downloads: z.array(
-    z.object({
-      exists: z.boolean().nullable(),
-      type: z.enum([
-        "presentation",
-        "intro-quiz-questions",
-        "intro-quiz-answers",
-        "exit-quiz-questions",
-        "exit-quiz-answers",
-        "worksheet-pdf",
-        "worksheet-pptx",
-        "supplementary-pdf",
-        "supplementary-pptx",
-      ]),
-      label: z.string(),
-      ext: z.string(),
-      forbidden: z.union([
-        z.array(z.object({ copyright_info: z.string() })),
-        z.boolean().optional().nullish(),
-      ]),
-    }),
-  ),
-  lessonSlug: z.string(),
-  lessonTitle: z.string(),
-  pathways: z.array(pathwaySchema),
+export const lessonDownloadsCanonicalSchema = baseLessonDownloadsSchema.extend({
+  pathways: z.array(lessonPathwaySchema),
 });
 
 export type LessonDownloadsCanonical = z.infer<
